refactor: workspace -> myspace로 변경
