+++ src/components/Myspace/SearchBar.tsx
