+++ src/pages/Myspace.tsx
@@ -1,0 +1,11 @@
+import Toolbar from '@/components/Myspace/Toolbar';
+
+const Myspace = () => {
+  return (
+    <div>
+      <Toolbar />
+    </div>
+  );
+};
+
+export default Myspace;
