--- src/pages/Workspace.tsx
@@ -1,11 +1,0 @@
-import Toolbar from '@/components/Workspace/Toolbar';
-
-const Workspace = () => {
-  return (
-    <div>
-      <Toolbar />
-    </div>
-  );
-};
-
-export default Workspace;
